Use correct path calculating function
My original function gave correct answers up to 4, which happens to be
the longest streak of distances of 1 jolt in the input. However I had
originally mistakenly calculated the number of possible paths through a
streak of 5 1s to be 11, where it's actually 13. This does not change
the answer, but it does change the correct function.

However I decided not to implement a Tribonacci sequence function in
full, and instead opted for a hardcoded version that runs faster and is
simpler to read. The information about the sequence itself was put in
the comments.